add components for maplibre popups and zonal statistics
